+++ packages/cater-emotion/server/layout.js
@@ -1,16 +1,16 @@
 // Copyright Jon Williams 2017. See LICENSE file.
-import ParentLayout, { Body, Scripts } from "server/^";
-import React from "react";
 import { extractCritical } from "emotion-server";
+import React from "react";
 import { renderToString } from "react-dom/server";
+import ParentLayout from 'server/^'; // Note: Inheritance import
+import Scripts from 'app/scripts';
 
 /**
  * Your regular Layout component, but with the emotion styles added in.
  */
 class Layout extends ParentLayout {
   render() {
-    const App = this.props.app;
-    const appBody = renderToString(<App />);
+    const appBody = renderToString(this.props.children);
     const { html, ids, css } = extractCritical(appBody);
     return (
       <html>
@@ -20,7 +20,7 @@ class Layout extends ParentLayout {
           <style dangerouslySetInnerHTML={{ __html: css }} />
         </head>
         <body>
-          <div id="root" dangerouslySetInnerHTML={{ __html: html }} />
+          {this.props.children}
           <Scripts />
         </body>
       </html>