+++ packages/cater/client/global-style.js
@@ -1,0 +1,9 @@
+// Copyright Jon Williams 2017. See LICENSE file.
+import Empty from "app/cater/empty";
+import React from "react";
+
+// Global styles are rendered on the server-side, so we have a stubbed
+// out component here.
+class GlobalStyle extends Empty {}
+
+export default GlobalStyle;