+++ packages/cater/src/context-options.js
@@ -37,4 +37,4 @@ DEFAULT_OPTIONS.babel = {
   ]
 };
 
-module.exports = DEFAULT_OPTIONS; // CommonJS
+module.exports = DEFAULT_OPTIONS;