+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - The caret import. This will import from a component that overrides another component. For example if you had project/server/layout import '^' it would return the cater/server/layout component. Useful for wrappering.
-- Asset processing from the project/asset directory. Supports images, scss and css files.
+- Asset processing from the project/asset directory. Supports images, scss and css files. Uses Babel transforms on the server-side and Webpack on the client side.
+- Moved to a plugin model that'll allow the base Cater code to be simpler.
 
 ### Changed
 - Updated dependent packages, including Jest from v21 to v22.